Moves Generate Plot to Plot Manager
